requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
pillow>=9.0.0
python-multipart>=0.0.6
numpy>=1.23.0
